improve qunit tests performance
